# perf(pnp): cache result of findApiPathFor