--- a/packages/yarnpkg-pnp/sources/loader/module.d.ts
+++ b/packages/yarnpkg-pnp/sources/loader/module.d.ts
@@ -10,6 +10,6 @@ declare module "module" {
   function _load(request: string, parent: NodeModule | null | undefined, isMain: boolean): any;
 
   interface Module extends NodeModule {
-    pnpApiPath?: import('@yarnpkg/fslib').PortablePath;
+    pnpApiPath?: import('@yarnpkg/fslib').PortablePath | null;
   }
 }