**What's the problem this PR addresses?**

When calling `getApiPathFromParent` we check if the module already has the `pnpApiPath`, if it doesn't we run `findApiPathFor` to find it. However we don't save the result to the module, which causes us to look for the api path multiple times for the same modules.

Running `yarn` on the repo shows this improvement:
||before|after|
|--|---|--|
|existsSync|12724 calls | 10996 calls (-1728)|
|statSync|15737 calls | 15545 calls (-192)|
|Time spent searching|145.72ms| 2.98ms|

Note: Testing on a much smaller repo the improvement to `Time spent searching` is ~3ms, from ~5ms to ~2ms


**How did you fix it?**

Store the result of `findApiPathFor` on the parent module

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.